Change how TextBoxFrame's border argument work
Before a default border was drawn when no argument was sent to the
border argument. Now no border is drawn when there is no argument.

The default border dict was instead moved to the borders.py module
and can now be used as other included borders.

This change was made because it should be possible to create a
dialog box without borders.